checkbox-touch: add largeBuffer flag to SelectionPage
UbuntuListView uses fancy logic to load and unload listitems when user drags
them into or out of view. That helps the performance, but prevents autopilot
proxy classes from seeing all the list items. The component has a cacheBuffer
property that gives a little leeway for the items to be kept in the memory (or
loaded in the first place). On the platforms I checked it's 320px.

This patch introduces a new property for the SelectionPage: 'largeBuffer',
that, when used, makes the cacheBuffer 10x larger which *should* make all the
items stay in the memory.

Note that this property should be used only for testing purposes as it may
reduce the overall performance. YMMV

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>