checkbox-touch: handle failed object creation in PythonObjectRef
When creation method call to python failed or misbehaved pyotherside sent in
None as the return value. This was treated as a valid response making
triggering of objectReady signal happen. The worst scenario for this is
PythonLogger construction failing resulting in writes to console.error which in
turn call the 'error' method of the object that is suposed to be a valid python
object, but is None instead.

This patch checks if the creation method returned None (or got None because of
python failing) and prevents the objectReady signal from being fired.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>