plainbox:session:state: add SessionDeviceContext.get_unit()
This patch adds an O(1) method for looking up a unit with a given
identifier in the session device context (so a unit that exists on a
given device.

This is done by tracking the addition and removal of appropriate units
and using a lookaside map that maintains the up-to-date state.

This solves the annoying problem that many apps solve by doing this
themselves or by computing a less effective solution.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>