plainbox:transport: add certification transport
This patch moves a (slightly improved) copy of the certification
transport from checkbox-ng over to plainbox. The reason for this is that
transports are not units and applications cannot easily discover and
configure transports in a high-level way. For the session assistant I
wanted to have easy-to-understand methods that let anyone create a
transport without having to import a lot of things and know appropriate
URLs.

The code in checkbox-ng is left intact. I hope to remove it when
rewriting chceckbox-ng on top of the SessionAssistant class.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>