checkbox-support: correct test suite loader
This patch corrects a simple bug in the test suite loader that was previously
silently compensated by earlier versions of the standard library. The bug has
caused us to effectively import modules and packages from _inside_ the
checkbox_support package. For example, the module 'checkbox_support.dbus' was
imported simply as 'dbus'. This caused all kinds of understandable havoc.  The
fix relies on the second, optional argument to defaultTestLoader.discover()
which lets us specify that we want to load 'checkbox_support' from the parent
directory.

Fixes: https://launchpad.net/bugs/1483410
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>