plainbox: add support for relocatable providers
This patch adds support for making effectively relocatable providers. A
provider like that can be in any location an as long as PROVIDERPATH is
set correctly (or the provider is found on the implicit path) it can be
moved around entirely.

This is accomplished with two parts:

1) The provider definition file can now include the relocatable=True
   key. When this key is present all of the other directory variables
   are ignored and they are always computed based on the location of the
   file itself.

2) The provider manager has a new layout, relocatable, that generates a
   set of compatible directories. They are identical to
   provider-under-development layout.

The main consumer of this feature is the snappy-friendly,
snapcraft-build provider where we don't really know where the provider
(or the whole application) is really at runtime (various versions,
side-loading, etc.). With this trick each provider can be simply
installed to a dedicated directory.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>