checkbox-touch:autopilot: invoke autopilot from *the current* python3
This patch makes invocation of autopilot tests (through ./tests/autopilot/run)
to use python3 that is available in the current environment. The default file
that's used when invoking autopilot (`/usr/bin/autopilot`) is a generated file
with shebang pointing to the system wide python interpreter, which - when
cannot import plainbox (or worse, uses out-of-date one).
This resulted in _get_plainbox_qml_modules_path() method trying to find
Plainbox qml module in the `./lib/(...)` directory instead of the venv one.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>