checkbox-touch: move embedded providers stuff to CBTApp.__init__
This patch changes when and where info about embedded providers directory is
propagated from QML to the python side.
Previously this path was pushed when starting or resuming session. When using
session assistant we need to load providers earlier, before we start checking
whether the session is resumable, which is one step before resuming or starting
the session.
This also makes the code cleaner, as we don't need to push this variable in
multiple calls. Note that this 'variable' is quite constant throughout the
lifetime of the app, as its value is read once from the setting.json.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>