plainbox: work around regular expression changes in py3.5
This patch reduces the error messages now produced by Python 3.5 to
those of Python 3.4 and earlier. It is something I wish we didn't have
to do but it's the only way that works quickly, without converting all
of the doctests over to unit tests.

Here, in particular, the information where a parse error was found
within the regular expression is being discarded.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>