plainbox: work around subtle API change in py35
This patch makes us use sre_constants.LITERAL (an enum) rather than the
string 'literal' that was being used in all the earlier versions of
python. Without this change the code that parsed regular expressions to
check if they are a trivial string comparison operation would never
work which was causing a cascade failure of other tests.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>