plainbox.ctrl: Force environment into non-internationalized values
When executing a job using plainbox-trusted-launcher, pkexec or root, a
"delta environment" is generated in CheckBoxDifferentialExecutionController
class.

If the job does not explicitely preserves the system locales, we set them
to use non-internationalized values.

Previously, keys like LANGUAGE or LC_ALL were not set in the differential
environment, resulting in plainbox using the one available system-wise. When
testing on non-English devices, this resulted in some jobs failing (when
expecting an English return value for instance).