checkbox-touch: fix horrible mem usage when job prints stuff
CommandOutput page used a buffer that was filled with the strings that job's
command printed and periodically timer synced that output onto the text area.
As the buffer qml's "property string" was used. This somehow broke garbage
collecting and the strings were never freed from memory. By changing the type
to 'property var' fixed all problems.
Why?
I don't know.
Here: http://paste.ubuntu.com/12310269/ is the code I did to pinpoint where/how
the problem may be solved. If you want to reproduce the bug, in the pastebinned
code change the type of bufferedIO from 'var' to 'string'.

Fixes: https://bugs.launchpad.net/checkbox-converged/+bug/1490533

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>