checkbox-touch: add a way for custom submission input dialogs
This patch adds a way how submission details can be supplied. If submission
requires sophisticated logic or a multi-step process, then simple sequence of
input fields is not enough. From know on, submission entry in settings can use
'inputForm' field that specifies which QML file to use as a GUI for submission
details dialog.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>