plainbox: add new job flag (has-leftovers)
This patch adds a new flag 'has-leftovers' available to all job
definitions that have a command. This flag that is especially made to
address a problem discovered by snapcraft (which is using plainbox)
where a provider is mostly filled with jobs that create a lot of files
as a course of normal operation.

Since plainbox normally logs a warning (controllable via API but not on
command line) whenever that happens the multitude of warnings made it
difficult to see what was going on in the actual tests. Test designers
can now use this flag to silence such messages on a case-by-case basis.

Fixes: https://bugs.launchpad.net/snapcraft/+bug/1484596
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>