checkbox-touch: fix latching of 'start testing' btn on welcome page
This patch fixes a problem where "Start testing" button on the welcome page
would be unlatched even when pressed.
It wasn't noticable on the first run, but when running any next session,
plainbox stack does some magic under the bonnet, and the time delay was
noticible. During that time the button was _pressable_ which enabled user to
spam the app with calls to get_testplans().

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>