provider:checkbox: Fix Wireless tests results regardless of Eth connection
When testing jobs such as `wireless/wireless_connection_*` or
`suspend/wireless_connection_after_suspend_*`, if an Ethernet cable was
plugged to the device being tested, it would bias the results and
provide erroneous feedback.

Now, before testing wireless connections, we make sure only the connection
we want to test is activated, and when the test is over, if needed, we
reactivate the Ethernet connection.