checkbox-touch: ignore resumable sessions having no app_blob
As we're trying to blindly deserialize app_blob from the last valid incomplete
session, we need to make sure that we actually can do this. This patch adds
checking if the items returned as resumable sessions by session assistant
contain app_blob.

This patch doesn't change anything in the session assistant, as a session
checkpoint without app_blob is not invalid.

Fixes: https://bugs.launchpad.net/checkbox-converged/+bug/1501500

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>