checkbox-touch: add logic for dropping permissions for confined tests
This patch adds the logic that pulls information from the test's qml file about
which permissions to clear, and calls appropriate bits.
To extract this info we first instantiate the QmlJob object in a fake container
Item, so te field is initiated as test author specified, then immedietaly (and
synchronously) destroyed, so all resources are (not sure about that :)) freed.
Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>