providers:qml-tests:camera-06: trigger summary on mediaplayer status
This patch fixes the camera test by changing what triggers summary screen to
pop up. Previously it was triggered by onStopped signal, but it seems this can
happen while mediaplayer is asked to play the video while it's still being
saved to file by the camera. So now we wait for the mediaplayer to have a
status of 'loaded' file, then we subscribe to the change of playbackState and
then we show the summary screen.

Fixes: https://bugs.launchpad.net/checkbox-converged/+bug/1492234

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>