checkbox-touch:components: support disabling of items in selection
This patch brings support to a new field for models in selection -
'mod_disabled', if set to true, makes item being rendered as disabled (no
reaction to clicks, a greyed-out style); it also makes '[de]select all' actions
not affect that item.

The field marks fact of being disabled instead of 'enabled' as it will be
treated as false if the field is not present, greatly reducing complexity of
the code.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>