checkbox-support:parsers:udevadm: Categorize eMMC internal drives as DISK
eMMC drives and MMC removable storage are exactly the same thing and it's
impossible to distinguish them.
But some recent systems have eMMC internal drives as their main and unique
storage device. In that case we need a way to mark them as DISK.

To do so, the idea is to check two things:

1. The / mount point will tell if we have to consider the device,
2. If the MMC type is actually "MMC" as we don't want to run heavy tests on SD
   cards (e.g on a pandaboard).

In order to get the mount points of all partitions, we're calling an external
command, `lsblk` (as libudev does not call libmount to provide such info).

The exact command is:

lsblk -i -n -P -o KNAME,TYPE,MOUNTPOINT

It works even on Ubuntu 12.04 (util-linux 2.20.1) and snappy.

The result of lsblk is a new parameter of UdevadmParser.