plainbox:transport: fix FileTransport for binary files
This patch removes assumption that the data coming to the send method of a
transport is a text. The target file was opened as a text file and (what's even
more stupid) ByteStringStreamTranslator was used to translate it to utf-8,
which, when used for exporters that create binaries (like xlsx one) failed
badly.

This patch removes the assume-it's-a-text code and treats data as is (Bytes).

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>