checkbox-ng:launchers: add retry submission option
This patch makes checkbox-cli handle failed submissions (caused by
TransportError) by displaying an error message and proposying a retry instead
of stacktracing and failing the whole session.

If some part of transport is user-supplied (think secure_id here), the user
will be asked to enter it again.

If the session is not interactive, checkbox will proceed to produce other
reports.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>