plainbox:exporter:tar: New exporter creating a tar.xz output
The archive contains two session exports (json and xlsx) and all test I/O logs.

Note that the job.stdout from the session/io-log is renamed to be just job in
the archive, but job.stderr is kept as is. It will be useful as some attachment
jobs are named like foo.jpg. that way it avoids getting a file named
foo.jpg.stdout.

Note #2, the I/O log filename is the slugified version of the job id.