checkbox-touch: fix freeze when going back to welcome page
So the crash happened because SessionAssistant wasn't expecting to get
get_test_plans calls when we're negotiating which tests to run.

There were handful of ways to fix it, namely:

1) Rebooting the session (finalizing, starting one afresh), this would require
maintaining information whether the session was already started. This would be
expensive to run (starting session). Note that we do this if the user changes
their mind about which test plan to run, as then we need to bootstrap
everything.

2) Letting get_test_plans be called in different stages of the session's
maturation process. This would bring allowed_calls extension in couple of
places, making SA code uglier and less linear in terms of session lifetime.

3) Remembering available test plans in the app state. Test plans don't change
once session is started, so this is safe. Note that once test plan is selected
the session will be rebooted.

I went with 3). Least code and least collateral additions.

Fixes: https://bugs.launchpad.net/checkbox-converged/+bug/1585030

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>