checkbox-touch: Add the launcher optional argument
The launcher argument is just the full pathname to the launcher definition file.
The same file used and processed by checkbox-cli.
If this argument is present, it takes precedence over the settings option.

The settings argument is still present but not required thus builds to create
a click package from checkbox-converged are not impacted to the launcher addition.