checkbox-touch: Submit results to C3 using the launcher settings
checkbox-cli does the launcher reports processing in a single step whereas
checkbox-converged do the same in two possible steps, saving reports locally
(optional though) and the submission step.

To keep those 2 actions separated and reuse most of the existing code, the
submission step cannot benefit of the report previously saved on disk.
After a call to get all the certification transport config from the launcher
settings, a full appSettings["submission"] is created and sent to the results
qml page.

One added value is that the secure_id popup is not triggered if the id is
provided in the launcher (using [transport:c3]).