plainbox:session:assistant: remove forgetting running_job_name
Assistant in various places records the name (id, really) of the job being run
in the session state metadata. This could be used to inform the user which job
would be resumed, but somehow, before running checkpoint() (that saves the
state) the running_job_name field was set to None.

This patch removes this resetting to None action.

I have no idea why it was there.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>