converged:autopilot: improve on swiping to get proper listitem
Previously the code just swiped to the bottom, and then swiped the screen, so
the listitem was in the view. The latter action could fail if the item was
already optimised-out.

This patch changes the algorithm to swipe down bit by bit, until the item is
found. This way it's always present.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>