plainbox:ctrl: Downvote the pkexec based controllers on snappy
When plainbox needs to run a job as root it will use a set of execution
controllers and elect the most appropriate for the task. Currently, that will
include the policykit-based controller that simply doesn't work on snappy. When
snappy environment is detected we should not consider that controller.

Fixes https://bugs.launchpad.net/plainbox/+bug/1473941