plainbox:unit: make validation work on validator instances ONLY
Previously field_validators of any unit could be a IFieldValidator instance, a
IFieldValidator descendant, or a list containg any of the previous two.
This was unintuitive, and as measurements shown, subotimal.

This patch replaces all that with instances.

It introduces concrete_validators module that contains commonly used
validators, so callees don't have to re-instantiate them.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>